# api-documentation.yaml
openapi: 3.0.0
info:
  title: PROgress Tracker API
  description: Breast Cancer Patient Survey Analytics Platform API
  version: 1.0.0
  contact:
    name: Support Team
    email: support@progresstracker.ca
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.progresstracker.ca
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Success:
      type: object
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
        data:
          type: object
...
paths:
...
  /api/upload/single:
    post:
      tags:
        - Upload
      summary: Upload single file
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      fileName:
                        type: string
                      recordsProcessed:
                        type: integer
                      insights:
                        type: array
                        items:
                          type: string
  /api/upload/multiple:
    post:
      tags:
        - Upload
      summary: Upload multiple files
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        200:
          description: Files uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /api/upload/history:
    get:
      tags:
        - Upload
      summary: Get upload history
      security:
        - bearerAuth: []
      responses:
        200:
          description: Upload history retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  uploads:
                    type: array
                    items:
                      type: object

  # Analytics Endpoints
...